pyproject.toml
[project]
name = "skymp-masterapi"
version = "0.0.1"
description = "REST API application for monitoring SkyMp servers and authorizing clients for these servers"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "celery>=5.5.3",
    "cryptography>=45.0.7",
    "fastapi>=0.116.2",
    "flower>=2.0.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "mimesis>=18.0.0",
    "orjson>=3.11.3",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.9",
    "pyjwt>=2.10.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-multipart>=0.0.20",
    "python-ulid[pydantic]>=3.1.0",
    "redis[hiredis]>=6.4.0",
    "ruff>=0.13.0",
    "sqlalchemy>=2.0.43",
    "toml>=0.10.2",
    "ua-parser>=1.0.1",
    "user-agents>=2.2.0",
    "uvicorn[standard]>=0.35.0",
]

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]
pythonpath = [".", "backend/"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
